Time: 2 ms (16.22%) | Memory: 69.4 MB (6.41%) - LeetSync

+++ 209-minimum-size-subarray-sum/minimum-size-subarray-sum.java
@@ -2,14 +2,12 @@ class Solution {
     public int minSubArrayLen(int target, int[] nums) {
         int n=nums.length;
         int i=0,j=0;
-        long sum=0;
+        int sum=0;
         int ans=Integer.MAX_VALUE;
 
 
         while(j<n){
             sum+=nums[j];
-
-
             while(i<=j && sum>=target){
                 ans=Math.min(ans,j-i+1);
